First iteration at refactoring MicroClimate
