+++ Models/MicroClimate/CanopyType.cs
@@ -1,55 +1,52 @@
-﻿using System;
-using Models.Interfaces;
-
+﻿
 namespace Models
 {
-    public partial class MicroClimate
-    {
-        /// <summary>Wraps a canopy object.</summary>
-        [Serializable]
-        private class CanopyType
-        {
-            /// <summary>The canopy.</summary>
-            public ICanopy Canopy;
-            /// <summary>The ktot</summary>
-            public double Ktot;
-            /// <summary>The k</summary>
-            public double K;
-            /// <summary>The layer lai</summary>
-            public double[] LAI;
-            /// <summary>The layer la itot</summary>
-            public double[] LAItot;
-            /// <summary>The ftot</summary>
-            public double[] Ftot;
-            /// <summary>The fgreen</summary>
-            public double[] Fgreen;
-            /// <summary>The rs</summary>
-            public double[] Rs;
-            /// <summary>The rl</summary>
-            public double[] Rl;
-            /// <summary>The rsoil</summary>
-            public double[] Rsoil;
-            /// <summary>The gc</summary>
-            public double[] Gc;
-            /// <summary>The ga</summary>
-            public double[] Ga;
-            /// <summary>The pet</summary>
-            public double[] PET;
-            /// <summary>The pe tr</summary>
-            public double[] PETr;
-            /// <summary>The pe ta</summary>
-            public double[] PETa;
-            /// <summary>The omega</summary>
-            public double[] Omega;
-            /// <summary>The interception</summary>
-            public double[] interception;
-            /// <summary>Constructor</summary>
-            /// <param name="canopy">The canopy to wrap.</param>
-            public CanopyType(ICanopy canopy)
-            {
-                Canopy = canopy;
-            }
-        }
+    using System;
+    using Models.Interfaces;
 
+    /// <summary>Wraps a canopy object.</summary>
+    [Serializable]
+    public class CanopyType
+    {
+        /// <summary>The canopy.</summary>
+        public ICanopy Canopy;
+        /// <summary>The ktot</summary>
+        public double Ktot;
+        /// <summary>The k</summary>
+        public double K;
+        /// <summary>The layer lai</summary>
+        public double[] LAI;
+        /// <summary>The layer la itot</summary>
+        public double[] LAItot;
+        /// <summary>The ftot</summary>
+        public double[] Ftot;
+        /// <summary>The fgreen</summary>
+        public double[] Fgreen;
+        /// <summary>The rs</summary>
+        public double[] Rs;
+        /// <summary>The rl</summary>
+        public double[] Rl;
+        /// <summary>The rsoil</summary>
+        public double[] Rsoil;
+        /// <summary>The gc</summary>
+        public double[] Gc;
+        /// <summary>The ga</summary>
+        public double[] Ga;
+        /// <summary>The pet</summary>
+        public double[] PET;
+        /// <summary>The pe tr</summary>
+        public double[] PETr;
+        /// <summary>The pe ta</summary>
+        public double[] PETa;
+        /// <summary>The omega</summary>
+        public double[] Omega;
+        /// <summary>The interception</summary>
+        public double[] interception;
+        /// <summary>Constructor</summary>
+        /// <param name="canopy">The canopy to wrap.</param>
+        public CanopyType(ICanopy canopy)
+        {
+            Canopy = canopy;
+        }
     }
 }
