+++ Models/MicroClimate/MicroClimate.cs
@@ -1,12 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using APSIM.Shared.Utilities;
-using Models.Core;
-using Models.Interfaces;
-using System.Linq;
-
-namespace Models
+﻿namespace Models
 {
+    using System;
+    using System.Collections.Generic;
+    using APSIM.Shared.Utilities;
+    using Models.Core;
+    using Models.Interfaces;
+    using System.Linq;
+
     /// <summary>
     /// # [Name]
     /// The module MICROMET, described here, has been developed to allow the calculation of 
@@ -17,98 +17,32 @@ namespace Models
     [PresenterName("UserInterface.Presenters.PropertyPresenter")]
     [ValidParent(ParentType = typeof(Simulation))]
     [ValidParent(ParentType = typeof(Zone))]
-    public partial class MicroClimate : Model
+    public class MicroClimate : Model
     {
-
-        #region Constants
-        /// <summary>The SVP_ a Teten coefficient</summary>
-        private const double svp_A = 6.106;
-        /// <summary>The SVP_ b Teten coefficient</summary>
-        private const double svp_B = 17.27;
-        /// <summary>The SVP_ c Teten coefficient</summary> 
-        private const double svp_C = 237.3;
-        /// <summary>0 C in Kelvin (k)</summary>
-        private const double abs_temp = 273.16;
-        /// <summary>universal gas constant (J/mol/K)</summary>
-        private const double r_gas = 8.3143;
-        /// <summary>molecular weight water (kg/mol)</summary>
-        private const double mwh2o = 0.018016;
-        /// <summary>molecular weight air (kg/mol)</summary>
-        private const double mwair = 0.02897;
-        /// <summary>molecular fraction of water to air ()</summary>
-        private const double molef = mwh2o / mwair;
-        /// <summary>Specific heat of air at constant pressure (J/kg/K)</summary>
-        private const double Cp = 1010.0;
-        /// <summary>Stefan-Boltzman constant</summary>
-        private const double stef_boltz = 5.67E-08;
-        /// <summary>constant for cloud effect on longwave radiation</summary>
-        private const double c_cloud = 0.1;
-        /// <summary>convert degrees to radians</summary>
-        private const double Deg2Rad = Math.PI / 180.0;
-        /// <summary>Density of water (kg/m3)</summary>
-        private const double RhoW = 998.0;
-        /// <summary>weights vpd towards vpd at maximum temperature</summary>
-        private const double svp_fract = 0.66;
-        /// <summary>The sun set angle (degrees)</summary>
-        private const double SunSetAngle = 0.0;
-        /// <summary>The sun angle for net positive radiation (degrees)</summary>
-        private const double SunAngleNetPositiveRadiation = 15;
-        /// <summary>Convert hours to seconds</summary>
-        private const double hr2s = 60.0 * 60.0;
-        /// <summary>von Karman constant</summary>
-        private const double vonKarman = 0.41;
-        /// <summary>Canopy emissivity</summary>
-        private const double CanopyEmissivity = 0.96;
-        /// <summary>The soil_emissivity</summary>
-        private const double SoilEmissivity = 0.96;
-
-        // CERES Eo Model Constants
-        /// <summary>
-        /// Temperature below which eeq decreases (oC)
-        /// </summary>
-        private double min_crit_temp = 5.0;
-
-        /// <summary>
-        /// Temperature above which eeq increases (oC)
-        /// </summary>
-        private double max_crit_temp = 35.0;
-
-        /// <summary>
-        /// albedo at 100% green crop cover (0-1)
-        /// </summary>
-        private double max_albedo = 0.23;
-
-        /// <summary>
-        /// albedo at 100% residue cover (0-1)
-        /// </summary>
-        private double residue_albedo = 0.23;
-
-
-        #endregion
-
-
         /// <summary>The clock</summary>
         [Link]
-        private Clock Clock = null;
+        private Clock clock = null;
 
         /// <summary>The weather</summary>
         [Link]
         private IWeather weather = null;
-        
+
         /// <summary>The surface organic matter model</summary>
         [Link(IsOptional = true)]
         private ISurfaceOrganicMatter residues = null;
 
-        /// <summary>List of uptakes</summary>
-        private List<ZoneMicroClimate> zoneMicroClimates = new List<ZoneMicroClimate>();
+        /// <summary>The sun set angle (degrees)</summary>
+        private const double sunSetAngle = 0.0;
 
-        /// <summary>Constructor</summary>
-        public MicroClimate()
-        {
-        }
+        /// <summary>The sun angle for net positive radiation (degrees)</summary>
+        private const double sunAngleNetPositiveRadiation = 15;
+
+        /// <summary>List of uptakes</summary>
+        private List<MicroClimateZone> microClimatesZones = new List<MicroClimateZone>();
 
         /// <summary>This is the length of time within the day during which evaporation will take place</summary>
         private double dayLengthEvap;
+
         /// <summary>This is the length of time within the day during which the sun is above the horizon</summary>
         private double dayLengthLight;
 
@@ -148,18 +82,16 @@ namespace Models
         [Units("MJ/MJ")]
         public double SoilHeatFluxFraction { get; set; }
 
-        /// <summary>
-        /// Height of the tallest canopy.
-        /// </summary>
+        /// <summary>Height of the tallest canopy.</summary>
         [Units("mm")]
         public double CanopyHeight
         {
             get
             {
-                if (zoneMicroClimates.Sum(n => n.Canopies.Count) == 0)
+                if (microClimatesZones.Sum(n => n.Canopies.Count) == 0)
                     return 0;
                 else
-                    return zoneMicroClimates.Max(m => m.Canopies.Max(c => c.Canopy.Height));
+                    return microClimatesZones.Max(m => m.Canopies.Max(c => c.Canopy.Height));
             }
         }
 
@@ -183,9 +115,9 @@ namespace Models
         {
             get
             {
-                double[] values = new double[zoneMicroClimates.Count];
-                for (int i = 0; i < zoneMicroClimates.Count; i++)
-                    values[i] = zoneMicroClimates[i].SurfaceRs;
+                double[] values = new double[microClimatesZones.Count];
+                for (int i = 0; i < microClimatesZones.Count; i++)
+                    values[i] = microClimatesZones[i].SurfaceRs;
 
                 return values;
             }
@@ -196,7 +128,7 @@ namespace Models
         [Units("mm")]
         public double PrecipitationInterception
         {
-            get { return zoneMicroClimates[0].PrecipitationInterception; }
+            get { return microClimatesZones[0].PrecipitationInterception; }
         }
 
         /// <summary>Gets the amount of radiation intercepted by the canopy (MJ/m2).</summary>
@@ -204,38 +136,8 @@ namespace Models
         [Units("MJ/m^2")]
         public double RadiationInterception
         {
-            get { return zoneMicroClimates[0].RadiationInterception; }
+            get { return microClimatesZones[0].RadiationInterception; }
         }
-
-        /*        /// <summary>Gets the canopy conductance aggregated for all canopy components (mm).</summary>
-                [Description("Whole system canopy conductance")]
-                [Units("mm")]
-                public double CanopyConductance
-                {
-                    get
-                    {
-                        double totalConductance = 0.0;
-                        for (int i = 0; i <= zoneMicroClimates[0].numLayers - 1; i++)
-                            for (int j = 0; j <= zoneMicroClimates[0].Canopies.Count - 1; j++)
-                                totalConductance += zoneMicroClimates[0].Canopies[j].Gc[i];
-                        return totalConductance;
-                    }
-                }
-
-                /// <summary>Gets the aerodynamic conductance aggregated for all canopy components (mm).</summary>
-                [Description("Whole system aerodynamic conductance")]
-                [Units("mm")]
-                public double AerodynamicConductance
-                {
-                    get
-                    {
-                        double totalConductance = 0.0;
-                        for (int i = 0; i <= zoneMicroClimates[0].numLayers - 1; i++)
-                            for (int j = 0; j <= zoneMicroClimates[0].Canopies.Count - 1; j++)
-                                totalConductance += zoneMicroClimates[0].Canopies[j].Ga[i];
-                        return totalConductance;
-                    }
-                }*/
 
         /// <summary>Gets the total Penman-Monteith potential evapotranspiration (MJ/m2).</summary>
         [Description("Total Penman-Monteith potential evapotranspiration")]
@@ -250,7 +152,7 @@ namespace Models
         [Units("mm")]
         public double PetRadiationTerm
         {
-            get { return zoneMicroClimates[0].petr; }
+            get { return microClimatesZones[0].petr; }
         }
 
         /// <summary>Gets the aerodynamic term of for the Penman-Monteith PET (mm).</summary>
@@ -258,7 +160,7 @@ namespace Models
         [Units("mm")]
         public double PetAerodynamicTerm
         {
-            get { return zoneMicroClimates[0].peta; }
+            get { return microClimatesZones[0].peta; }
         }
 
         /// <summary>Gets the fraction of the daytime in which the leaves are dry (0-1).</summary>
@@ -266,7 +168,7 @@ namespace Models
         [Units("-")]
         public double DryLeafTimeFraction
         {
-            get { return zoneMicroClimates[0].DryLeafFraction; }
+            get { return microClimatesZones[0].DryLeafFraction; }
         }
 
         /// <summary>Gets the total net radiation, long and short waves (MJ/m2).</summary>
@@ -283,7 +185,7 @@ namespace Models
         [Units("MJ/m^2")]
         public double NetShortWaveRadiation
         {
-            get { return weather.Radn * (1.0 - zoneMicroClimates[0].Albedo); }
+            get { return weather.Radn * (1.0 - microClimatesZones[0].Albedo); }
         }
 
         /// <summary>Gets the net long wave radiation (MJ/m2).</summary>
@@ -291,7 +193,7 @@ namespace Models
         [Units("MJ/m^2")]
         public double NetLongWaveRadiation
         {
-            get { return zoneMicroClimates[0].NetLongWaveRadiation; }
+            get { return microClimatesZones[0].NetLongWaveRadiation; }
         }
 
         /// <summary>Gets the flux of heat into the soil (MJ/m2).</summary>
@@ -299,7 +201,7 @@ namespace Models
         [Units("MJ/m^2")]
         public double SoilHeatFlux
         {
-            get { return zoneMicroClimates[0].SoilHeatFlux; }
+            get { return microClimatesZones[0].SoilHeatFlux; }
         }
 
         /// <summary>Gets the total plant cover (0-1).</summary>
@@ -316,32 +218,11 @@ namespace Models
         [EventSubscribe("StartOfDay")]
         private void OnStartOfDay(object sender, EventArgs e)
         {
-            zoneMicroClimates.Clear();
+            microClimatesZones.Clear();
             foreach (Zone newZone in Apsim.ChildrenRecursively(this.Parent, typeof(Zone)))
-                CreateZoneMicroClimate(newZone);
-            if (zoneMicroClimates.Count == 0)
-                CreateZoneMicroClimate(this.Parent as Zone);
-        }
-
-        /// <summary>
-        /// Create a new MicroClimateZone for a given simulation zone
-        /// </summary>
-        /// <param name="newZone"></param>
-        private void CreateZoneMicroClimate(Zone newZone)
-        {
-            ZoneMicroClimate myZoneMC = new ZoneMicroClimate();
-            myZoneMC.zone = newZone;
-            myZoneMC.Reset();
-            foreach (ICanopy canopy in Apsim.ChildrenRecursively(newZone, typeof(ICanopy)))
-                if (canopy.Height > 0)
-                    myZoneMC.Canopies.Add(new CanopyType(canopy));
-            if (residues != null)
-                foreach (ICanopy residue in residues.ResidueLayers)
-                {
-                    myZoneMC.Canopies.Add(new CanopyType(residue));
-                }
-            zoneMicroClimates.Add(myZoneMC);
-
+                microClimatesZones.Add(new MicroClimateZone(clock, weather, newZone, residues));
+            if (microClimatesZones.Count == 0)
+                microClimatesZones.Add(new MicroClimateZone(clock, weather, this.Parent as Zone, residues));
         }
 
         /// <summary>Called when the canopy energy balance needs to be calculated.</summary>
@@ -350,77 +231,56 @@ namespace Models
         [EventSubscribe("DoEnergyArbitration")]
         private void DoEnergyArbitration(object sender, EventArgs e)
         {
-            dayLengthLight = MathUtilities.DayLength(Clock.Today.DayOfYear, SunSetAngle, weather.Latitude);
-            dayLengthEvap = MathUtilities.DayLength(Clock.Today.DayOfYear, SunAngleNetPositiveRadiation, weather.Latitude);
+            dayLengthLight = MathUtilities.DayLength(clock.Today.DayOfYear, sunSetAngle, weather.Latitude);
+            dayLengthEvap = MathUtilities.DayLength(clock.Today.DayOfYear, sunAngleNetPositiveRadiation, weather.Latitude);
             // VOS - a temporary kludge to get this running for high latitudes. MicroMet is due for a clean up soon so reconsider then.
             dayLengthEvap = Math.Max(dayLengthEvap, (dayLengthLight * 2.0 / 3.0));
-            if (zoneMicroClimates.Count == 2 && zoneMicroClimates[0].zone is Zones.RectangularZone 
-                && zoneMicroClimates[1].zone is Zones.RectangularZone && zoneMicroClimates[0].Canopies.Count > 0 && zoneMicroClimates[1].Canopies.Count > 0)
+            if (microClimatesZones.Count == 2 && microClimatesZones[0].Zone is Zones.RectangularZone 
+                && microClimatesZones[1].Zone is Zones.RectangularZone && microClimatesZones[0].Canopies.Count > 0 && microClimatesZones[1].Canopies.Count > 0)
             {
                 // We are in a strip crop simulation
-                zoneMicroClimates[0].DoCanopyCompartments();
-                zoneMicroClimates[1].DoCanopyCompartments();
+                microClimatesZones[0].DoCanopyCompartments();
+                microClimatesZones[1].DoCanopyCompartments();
                 CalculateStripZoneShortWaveRadiation();
             }
             else // Normal 1D zones are to be used
-                foreach (ZoneMicroClimate ZoneMC in zoneMicroClimates)
+                foreach (MicroClimateZone ZoneMC in microClimatesZones)
                 {
                     ZoneMC.DoCanopyCompartments();
                     CalculateLayeredShortWaveRadiation(ZoneMC, weather.Radn);
                 }
 
             // Light distribution is now complete so calculate remaining micromet equations
-            foreach (ZoneMicroClimate ZoneMC in zoneMicroClimates)
+            foreach (var ZoneMC in microClimatesZones)
             {
-                CalculateEnergyTerms(ZoneMC);
-                CalculateLongWaveRadiation(ZoneMC);
-                CalculateSoilHeatRadiation(ZoneMC);
-                CalculateGc(ZoneMC);
-                CalculateGa(ZoneMC);
-                CalculateInterception(ZoneMC);
-                CalculatePM(ZoneMC);
-                CalculateOmega(ZoneMC);
-                SetCanopyEnergyTerms(ZoneMC);
-                CalculateEo(ZoneMC);
+                ZoneMC.CalculateEnergyTerms(soil_albedo);
+                ZoneMC.CalculateLongWaveRadiation(dayLengthLight, dayLengthEvap);
+                ZoneMC.CalculateSoilHeatRadiation(SoilHeatFluxFraction);
+                ZoneMC.CalculateGc(dayLengthEvap);
+                ZoneMC.CalculateGa(ReferenceHeight);
+                ZoneMC.CalculateInterception(a_interception, b_interception, c_interception, d_interception);
+                ZoneMC.CalculatePM(dayLengthEvap, NightInterceptionFraction);
+                ZoneMC.CalculateOmega();
+                ZoneMC.SetCanopyEnergyTerms();
+                ZoneMC.CalculateEo();
             }
         }
 
-
-        /// <summary>Calculate the amtospheric potential evaporation rate for each zone</summary>
-        private void CalculateEo(ZoneMicroClimate ZoneMC)
-        {
-            ISoilWater zoneSoilWater = Apsim.Find(ZoneMC.zone, typeof(ISoilWater)) as ISoilWater;
-            ISurfaceOrganicMatter zoneSurfaceOM = Apsim.Find(ZoneMC.zone, typeof(ISurfaceOrganicMatter)) as ISurfaceOrganicMatter;
-
-            double CoverGreen = 0;
-            for (int j = 0; j <= ZoneMC.Canopies.Count - 1; j++)
-                if (ZoneMC.Canopies[j].Canopy != null)
-                   CoverGreen+= (1-CoverGreen)*ZoneMC.Canopies[j].Canopy.CoverGreen;
-
-            if (weather != null && zoneSoilWater != null && zoneSurfaceOM != null)
-               zoneSoilWater.Eo = AtmosphericPotentialEvaporationRate(weather.Radn, 
-                                                            weather.MaxT, 
-                                                            weather.MinT, 
-                                                            zoneSoilWater.Salb, 
-                                                            zoneSurfaceOM.Cover, 
-                                                            CoverGreen); 
-        }
-
-    ///<summary> Calculate the short wave radiation balance for strip crop system</summary>
-    private void CalculateStripZoneShortWaveRadiation()
+        ///<summary> Calculate the short wave radiation balance for strip crop system</summary>
+        private void CalculateStripZoneShortWaveRadiation()
         {
             
-            ZoneMicroClimate tallest;
-            ZoneMicroClimate shortest;
-            if (MathUtilities.Sum(zoneMicroClimates[0].DeltaZ)> MathUtilities.Sum(zoneMicroClimates[1].DeltaZ))
+            MicroClimateZone tallest;
+            MicroClimateZone shortest;
+            if (MathUtilities.Sum(microClimatesZones[0].DeltaZ)> MathUtilities.Sum(microClimatesZones[1].DeltaZ))
             {
-                tallest = zoneMicroClimates[0];
-                shortest = zoneMicroClimates[1];
+                tallest = microClimatesZones[0];
+                shortest = microClimatesZones[1];
             }
             else
             {
-                tallest = zoneMicroClimates[1];
-                shortest = zoneMicroClimates[0];
+                tallest = microClimatesZones[1];
+                shortest = microClimatesZones[0];
             }
 
             bool TallestIsTree = false;
@@ -440,7 +300,7 @@ namespace Models
         /// </summary>
         /// <param name="tree"></param>
         /// <param name="alley"></param>
-        private void doTreeRowCropShortWaveRadiation(ref ZoneMicroClimate tree, ref ZoneMicroClimate alley)
+        private void doTreeRowCropShortWaveRadiation(ref MicroClimateZone tree, ref MicroClimateZone alley)
         {
             if (MathUtilities.Sum(tree.DeltaZ) > 0)  // Don't perform calculations if layers are empty
             {
@@ -460,8 +320,8 @@ namespace Models
                 double Ha = MathUtilities.Sum(alley.DeltaZ);               // Height of alley canopy
                 if ((Ha > CBHt) & (tree.DeltaZ.Length > 1))
                     throw (new Exception("Height of the alley canopy must not exceed the base height of the tree canopy"));
-                double Wt = (tree.zone as Zones.RectangularZone).Width;    // Width of tree zone
-                double Wa = (alley.zone as Zones.RectangularZone).Width;   // Width of alley zone
+                double Wt = (tree.Zone as Zones.RectangularZone).Width;    // Width of tree zone
+                double Wa = (alley.Zone as Zones.RectangularZone).Width;   // Width of alley zone
                 double CWt = 0;//Math.Min(tree.Canopies[0].Canopy.Width / 1000, (Wt + Wa));// Width of the tree canopy
                 foreach (CanopyType c in tree.Canopies)
                     if (c.Canopy.Depth < c.Canopy.Height)
@@ -522,14 +382,14 @@ namespace Models
         /// </summary>
         /// <param name="tallest"></param>
         /// <param name="shortest"></param>
-        private void doStripCropShortWaveRadiation(ref ZoneMicroClimate tallest, ref ZoneMicroClimate shortest)
+        private void doStripCropShortWaveRadiation(ref MicroClimateZone tallest, ref MicroClimateZone shortest)
         {
             if (MathUtilities.Sum(tallest.DeltaZ) > 0)  // Don't perform calculations if layers are empty
             {
                 double Ht = MathUtilities.Sum(tallest.DeltaZ);                // Height of tallest strip
                 double Hs = MathUtilities.Sum(shortest.DeltaZ);               // Height of shortest strip
-                double Wt = (tallest.zone as Zones.RectangularZone).Width;    // Width of tallest strip
-                double Ws = (shortest.zone as Zones.RectangularZone).Width;   // Width of shortest strip
+                double Wt = (tallest.Zone as Zones.RectangularZone).Width;    // Width of tallest strip
+                double Ws = (shortest.Zone as Zones.RectangularZone).Width;   // Width of shortest strip
                 double Ft = Wt / (Wt + Ws);                                   // Fraction of space in tallest strip
                 double Fs = Ws / (Wt + Ws);                                   // Fraction of space in the shortest strip
                 double LAIt = MathUtilities.Sum(tallest.LAItotsum);           // LAI of tallest strip
@@ -576,134 +436,25 @@ namespace Models
 
 
         }
-    
 
-        /// <summary>Calculate the canopy conductance for system compartments</summary>
-        private void CalculateGc(ZoneMicroClimate ZoneMC)
+        /// <summary>
+        /// Calculates interception of short wave by canopy compartments
+        /// </summary>
+        private void CalculateLayeredShortWaveRadiation(MicroClimateZone ZoneMC, double Rin)
         {
-            double Rin = weather.Radn;
-
+            // Perform Top-Down Light Balance
+            // ==============================
+            double Rint = 0;
             for (int i = ZoneMC.numLayers - 1; i >= 0; i += -1)
             {
-                double Rflux = Rin * 1000000.0 / (dayLengthEvap * hr2s) * (1.0 - ZoneMC.Albedo);
-                double Rint = 0.0;
-
+                if (double.IsNaN(Rint))
+                    throw new Exception("Bad Radiation Value in Light partitioning");
+                Rint = Rin * (1.0 - Math.Exp(-ZoneMC.layerKtot[i] * ZoneMC.LAItotsum[i]));
                 for (int j = 0; j <= ZoneMC.Canopies.Count - 1; j++)
-                {
-                    ZoneMC.Canopies[j].Gc[i] = CropCanopyConductance(ZoneMC.Canopies[j].Canopy.Gsmax, ZoneMC.Canopies[j].Canopy.R50, ZoneMC.Canopies[j].Fgreen[i], ZoneMC.layerKtot[i], ZoneMC.LAItotsum[i], Rflux);
-                    Rint += ZoneMC.Canopies[j].Rs[i];
-                }
-                // Calculate Rin for the next layer down
+                    ZoneMC.Canopies[j].Rs[i] = Rint * MathUtilities.Divide(ZoneMC.Canopies[j].Ftot[i] * ZoneMC.Canopies[j].Ktot, ZoneMC.layerKtot[i], 0.0);
                 Rin -= Rint;
             }
-        }
-
-        /// <summary>Calculate the aerodynamic conductance for system compartments</summary>
-        private void CalculateGa(ZoneMicroClimate ZoneMC)
-        {
-            double sumDeltaZ = MathUtilities.Sum(ZoneMC.DeltaZ);
-            double sumLAI = MathUtilities.Sum(ZoneMC.LAItotsum);
-            double totalGa = AerodynamicConductanceFAO(weather.Wind, ReferenceHeight, sumDeltaZ, sumLAI);
-
-            for (int i = 0; i <= ZoneMC.numLayers - 1; i++)
-                for (int j = 0; j <= ZoneMC.Canopies.Count - 1; j++)
-                    ZoneMC.Canopies[j].Ga[i] = totalGa * MathUtilities.Divide(ZoneMC.Canopies[j].Rs[i], ZoneMC.sumRs, 0.0);
-        }
-
-        /// <summary>Calculate the interception loss of water from the canopy</summary>
-        private void CalculateInterception(ZoneMicroClimate ZoneMC)
-        {
-            double sumLAI = 0.0;
-            double sumLAItot = 0.0;
-            for (int i = 0; i <= ZoneMC.numLayers - 1; i++)
-                for (int j = 0; j <= ZoneMC.Canopies.Count - 1; j++)
-                {
-                    sumLAI += ZoneMC.Canopies[j].LAI[i];
-                    sumLAItot += ZoneMC.Canopies[j].LAItot[i];
-                }
-
-            double totalInterception = a_interception * Math.Pow(weather.Rain, b_interception) + c_interception * sumLAItot + d_interception;
-            totalInterception = Math.Max(0.0, Math.Min(0.99 * weather.Rain, totalInterception));
-
-            for (int i = 0; i <= ZoneMC.numLayers - 1; i++)
-                for (int j = 0; j <= ZoneMC.Canopies.Count - 1; j++)
-                    ZoneMC.Canopies[j].interception[i] = MathUtilities.Divide(ZoneMC.Canopies[j].LAI[i], sumLAI, 0.0) * totalInterception;
-
-            ISoilWater zonesoilwater = Apsim.Find(ZoneMC.zone, typeof(ISoilWater)) as ISoilWater;
-            if (zonesoilwater != null)
-            {
-                zonesoilwater.PotentialInfiltration = Math.Max(0, weather.Rain - totalInterception);
-                zonesoilwater.PrecipitationInterception = totalInterception;
-            }
-        }
-
-        /// <summary>Calculate the Penman-Monteith water demand</summary>
-        private void CalculatePM(ZoneMicroClimate ZoneMC)
-        {
-            // zero a few things, and sum a few others
-            double sumRl = 0.0;
-            double sumRsoil = 0.0;
-            double sumInterception = 0.0;
-            double freeEvapGa = 0.0;
-
-            for (int j = 0; j <= ZoneMC.Canopies.Count - 1; j++)
-            {
-                sumRl += MathUtilities.Sum(ZoneMC.Canopies[j].Rl);
-                sumRsoil += MathUtilities.Sum(ZoneMC.Canopies[j].Rsoil);
-                sumInterception += MathUtilities.Sum(ZoneMC.Canopies[j].interception);
-                freeEvapGa += MathUtilities.Sum(ZoneMC.Canopies[j].Ga);
-            }
-
-            double netRadiation = ((1.0 - ZoneMC.Albedo) * ZoneMC.sumRs + sumRl + sumRsoil) * 1000000.0;   // MJ/J
-            netRadiation = Math.Max(0.0, netRadiation);
-
-            double freeEvapGc = freeEvapGa * 1000000.0; // infinite surface conductance
-            double freeEvap = CalcPenmanMonteith(netRadiation, weather.MinT, weather.MaxT, weather.VP, weather.AirPressure, dayLengthEvap, freeEvapGa, freeEvapGc);
-
-            ZoneMC.DryLeafFraction = 1.0 - MathUtilities.Divide(sumInterception * (1.0 - NightInterceptionFraction), freeEvap, 0.0);
-            ZoneMC.DryLeafFraction = Math.Max(0.0, ZoneMC.DryLeafFraction);
-
-            for (int i = 0; i <= ZoneMC.numLayers - 1; i++)
-                for (int j = 0; j <= ZoneMC.Canopies.Count - 1; j++)
-                {
-                    netRadiation = 1000000.0 * ((1.0 - ZoneMC.Albedo) * ZoneMC.Canopies[j].Rs[i] + ZoneMC.Canopies[j].Rl[i] + ZoneMC.Canopies[j].Rsoil[i]);
-                    netRadiation = Math.Max(0.0, netRadiation);
-
-                    ZoneMC.Canopies[j].PETr[i] = CalcPETr(netRadiation * ZoneMC.DryLeafFraction, weather.MinT, weather.MaxT, weather.AirPressure, ZoneMC.Canopies[j].Ga[i], ZoneMC.Canopies[j].Gc[i]);
-                    ZoneMC.Canopies[j].PETa[i] = CalcPETa(weather.MinT, weather.MaxT, weather.VP, weather.AirPressure, dayLengthEvap * ZoneMC.DryLeafFraction, ZoneMC.Canopies[j].Ga[i], ZoneMC.Canopies[j].Gc[i]);
-                    ZoneMC.Canopies[j].PET[i] = ZoneMC.Canopies[j].PETr[i] + ZoneMC.Canopies[j].PETa[i];
-                }
-        }
-
-        /// <summary>Calculate the aerodynamic decoupling for system compartments</summary>
-        private void CalculateOmega(ZoneMicroClimate MCZone)
-        {
-            for (int i = 0; i <= MCZone.numLayers - 1; i++)
-                for (int j = 0; j <= MCZone.Canopies.Count - 1; j++)
-                    MCZone.Canopies[j].Omega[i] = CalcOmega(weather.MinT, weather.MaxT, weather.AirPressure, MCZone.Canopies[j].Ga[i], MCZone.Canopies[j].Gc[i]);
-        }
-
-        /// <summary>Send an energy balance event</summary>
-        private void SetCanopyEnergyTerms(ZoneMicroClimate ZoneMC)
-        {
-            for (int j = 0; j <= ZoneMC.Canopies.Count - 1; j++)
-                if (ZoneMC.Canopies[j].Canopy != null)
-                {
-                    CanopyEnergyBalanceInterceptionlayerType[] lightProfile = new CanopyEnergyBalanceInterceptionlayerType[ZoneMC.numLayers];
-                    double totalPotentialEp = 0;
-                    double totalInterception = 0.0;
-                    for (int i = 0; i <= ZoneMC.numLayers - 1; i++)
-                    {
-                        lightProfile[i] = new CanopyEnergyBalanceInterceptionlayerType();
-                        lightProfile[i].thickness = ZoneMC.DeltaZ[i];
-                        lightProfile[i].amount = ZoneMC.Canopies[j].Rs[i] * ZoneMC.RadnGreenFraction(j);
-                        totalPotentialEp += ZoneMC.Canopies[j].PET[i];
-                        totalInterception += ZoneMC.Canopies[j].interception[i];
-                    }
-                    ZoneMC.Canopies[j].Canopy.PotentialEP = totalPotentialEp;
-                    ZoneMC.Canopies[j].Canopy.WaterDemand = totalPotentialEp;
-                    ZoneMC.Canopies[j].Canopy.LightProfile = lightProfile;
-                }
+            ZoneMC.SurfaceRs = Rin;
         }
     }
 }
